RaskellDownload

Description

Simple package to allow haskell module downloads in raskell.

Packages

data Package

The Package type describes all information about a raskell-git-download package.

packageRepository describes the repository.

rootDir specifies in which location in the iOs Raskell filesystem the package will be deployed.

modules is a list of haskell modules to download.

Constructors

Package

Fields

packageRepository :: Repository

rootDir :: String

modules :: [[String]]

downloadPackage :: Package -> IO ()

downloadPackage p downloads a packge.

Repositories

data Repository

Data type to describe repository information.

repository contains the github username and the repo name.

prefix is usually "src/" or something similar.

branch allows to specify a branch such as master or develop

Constructors

Repository

Fields

repository :: String

prefix :: String

branch :: String

raskellGitDownload Package

raskellGitDownload :: Package

description of raskellGitDownload package

installRaskellGitDownload :: IO ()

installRaskellGitDownload install the installRaskellGitDownload