RaskellDownload

Description

Simple package to allow haskell module downloads in raskell.

Packages

data Package

The Package type describes all information about a raskell-git-download package.

packageRepository

packageRepository describes the repository.

rootDir

rootDir specifies in which location in the iOs Raskell filesystem the package will be deployed.

modules

modules is a list of haskell modules to download.

Constructors

Package

Fields

packageRepository :: Repository

rootDir :: String

modules :: [[String]]

downloadPackage :: Package -> IO ()

downloadPackage p downloads a packge.

Repositories

data Repository

Data type to describe repository information.

owner

owner specifies the github username

repository

repository contains the repo name.

authToken

you can supply an authentication token for a repository that will be used by the path api to autheticate access to the repository

prefix

prefix is usually "src/" or something similar.

branch

branch allows to specify a branch such as master or develop

pathApi

pathApi let's you select an API for repository paths the default API is for github (githubApiRaw) other apis available in this module are gogsApiRaw and githubApiV3

Constructors

Repository

Fields

owner :: String

repository :: String

authToken :: Maybe String

prefix :: String

branch :: String

pathApi :: Maybe PathApi

Path Apis

data PathApi

Data Type for Path APIs.

rawUrl

if we have an api :: PathApi then the expression downloadUrl = rawUrl api owner repo path branch token evaluates to the URL from which the content of path in owner's repository repo can be downloaded.

toRawContents

if fileContents == fetch downloadUrl then toRawContents api fileContents yield the actual file contents from fileContents as a string.

Constructors

PathApi

Fields

rawUrl :: String -> String -> String -> Maybe String -> Maybe String -> String

toRawContents :: String -> String

githubApiRaw :: PathApi

Path API for github server (https://github.com/)

gogsApiRaw :: [Char] -> PathApi

Path API for gogs git server (https://gogs.io/)

githubApiV3 :: PathApi

Path API for github server using the github API v3 (https://developer.github.com/v3/)

raskellGitDownload Package

raskellGitDownload :: Package

description of raskellGitDownload package

installRaskellGitDownload :: IO ()

installRaskellGitDownload install the installRaskellGitDownload